Refactoring and mapping participant
